attempted to have colliders for every tile, but realized there was tileMapColliders2D.
Need to know how to have trigger's display on console. Rn, the grass_Interactive tilemap has colliders that I wish the console displays "triggered" when the lawnmower moves over them, but I think the colliders are only at the edges of the tiles. Wanna see if the tiles can be individually placed (potentially) and interacted with or, if that's not possible, do it with the Grass_interative tilemap.